Initial experimental extension API
This extension API exists to be an experimental placeholder that
can be used by molecula to talk to extensions, allowing extensions
to be built using only this API, and thus, not requiring them to
depend on or import the rest of pilosa. It's deprecated; we already
have a new design in the works, but this placeholder lets us get
things done in the mean time.